# Use zero-copy C data interface in Parquet adapter